requirements.txt
asgiref==3.2.7
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
distro==1.5.0
Django==3.0.7
djangorestframework==3.11.0
docker==4.2.1
docker-compose==1.26.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
paramiko==2.7.1
Pillow==7.1.2
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
uritemplate==3.0.1
urllib3==1.25.9
websocket-client==0.57.0
zipp==3.1.0
